feat: extract process-cv route
